Revert "Merge pull request #300 from sopt-makers/feature/#263"
This reverts commit fa65c290153c464cb9ebca129a9378846a18b28f.
