+++ apiHooks/members.ts
@@ -1,14 +1,52 @@
 import { AxiosError } from 'axios';
-import { useMutation, useQuery, UseQueryOptions } from 'react-query';
+import { QueryKey, useInfiniteQuery, useMutation, useQuery, UseQueryOptions } from 'react-query';
 
 import {
   getMemberOfMe,
+  getMemberProfile,
   getMemberProfileById,
   getMemberProfileOfMe,
   getMembersSearchByName,
   postMemberCoffeeChat,
 } from '@/api/members';
-import { PostMemberCoffeeChatVariables, ProfileDetail } from '@/api/members/type';
+import { PostMemberCoffeeChatVariables, Profile, ProfileDetail } from '@/api/members/type';
+
+interface Variables {
+  limit?: number;
+  queryKey?: QueryKey;
+}
+
+// 멤버 프로필 전체 조회
+export const useGetMemberProfile = ({ limit, queryKey }: Variables) => {
+  const _queryKey = (typeof queryKey === 'string' ? [queryKey] : queryKey) ?? [];
+  return useInfiniteQuery({
+    queryKey: ['getMemberProfile', limit, ..._queryKey],
+    queryFn: async ({ pageParam: cursor = 0 }) => {
+      const params = { limit, cursor };
+
+      const apiUrl = new URL(window.location.href);
+      Object.entries(params).forEach(
+        ([query, value]) => value !== undefined && apiUrl.searchParams.set(query, value.toString()),
+      );
+
+      const data = await getMemberProfile(apiUrl.search);
+      return data;
+    },
+    getNextPageParam: (lastPage: Profile[]) => {
+      // TODO(@jun): nextPage 있는지 여부 boolean으로 undefined 예외처리
+      if (!lastPage.length) {
+        return undefined;
+      }
+      const lastIndex = lastPage.length - 1;
+      const lastMemberId = lastPage[lastIndex].id;
+      return lastMemberId;
+    },
+    onError: (error: { message: string }) => {
+      console.error(error.message);
+    },
+    keepPreviousData: true,
+  });
+};
 
 // 멤버 프로필 조회
 export const useGetMemberOfMe = () => {
