+++ package.json
@@ -36,7 +36,6 @@
     "lodash-es": "^4.17.21",
     "nanoid": "^4.0.0",
     "next": "^13.1.1",
-    "qs": "^6.11.0",
     "react": "^18.2.0",
     "react-datepicker": "^4.8.0",
     "react-dom": "^18.2.0",
